+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Convertkit React Changelog
 
-### [v1.0.8](https://github.com/ConvertKit/convertkit-react/releases/tag/v1.0.8) - (not released)
+### [v1.0.9](https://github.com/ConvertKit/convertkit-react/releases/tag/v1.0.9)
+
+- Change default background image of Rainer inline template
+
+### [v1.0.8](https://github.com/ConvertKit/convertkit-react/releases/tag/v1.0.8)
 
 - Add partial support for Mills and Rainer inline templates
 - Change default submit text to "Subscribe"